% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims_elig.R
\name{claims_elig}
\alias{claims_elig}
\title{Claims data member eligibility and demographics}
\usage{
claims_elig(conn, source = c("apcd", "mcaid", "mcaid_mcare", "mcare"),
  from_date = Sys.Date() - months(18), to_date = Sys.Date() -
  months(6), cov_min = 0, covgap_max = NULL, dual_min = NULL,
  dual_max = NULL, med_covgrp = NULL, pharm_covgrp = NULL,
  bsp_group_name = NULL, full_benefit_min = NULL, cov_type = NULL,
  mco_id = NULL, part_a_min = NULL, part_a_max = NULL,
  part_b_min = NULL, part_b_max = NULL, part_c_min = NULL,
  part_c_max = NULL, buy_in_min = NULL, buy_in_max = NULL,
  id = NULL, age_min = NULL, age_max = NULL, female = NULL,
  male = NULL, gender_me = NULL, gender_recent = NULL,
  race_aian = NULL, race_asian = NULL, race_asian_pi = NULL,
  race_black = NULL, race_nhpi = NULL, race_latino = NULL,
  race_white = NULL, race_unk = NULL, race_me = NULL,
  race_eth_me = NULL, race_recent = NULL, race_eth_recent = NULL,
  lang_amharic = NULL, lang_arabic = NULL, lang_chinese = NULL,
  lang_english = NULL, lang_korean = NULL, lang_russian = NULL,
  lang_somali = NULL, lang_spanish = NULL, lang_ukrainian = NULL,
  lang_vietnamese = NULL, lang_me = NULL, lang_recent = NULL,
  geo_zip = NULL, geo_hra_code = NULL, geo_school_code = NULL,
  geo_region = NULL, geo_county_code = NULL, geo_ach_code = NULL,
  geo_kc_ever = NULL, geo_kc_min = NULL,
  timevar_denom = c("duration", "cov_days"), show_query = T)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package}

\item{source}{Which claims data source do you want to pull from?}

\item{from_date}{Begin date for coverage period, "YYYY-MM-DD", 
defaults to 18 months prior to today's date (A)}

\item{to_date}{End date for coverage period, "YYYY-MM-DD", 
defaults to 6 months prior to today's date (A)}

\item{cov_min}{Minimum coverage required during requested date range (percent scale), defaults to 0 (A)}

\item{covgap_max}{Maximum gap in continuous coverage allowed during requested date range (days) (A)}

\item{dual_min}{Minimum Medicare-Medicaid dual eligibility coverage allowed 
during requested date range (percent scale) (A)}

\item{dual_max}{Maximum Medicare-Medicaid dual eligibility coverage allowed 
during requested date range (percent scale) (A)}

\item{med_covgrp}{Medical coverage group type in APCD data (AP)}

\item{pharm_covgrp}{Pharmacy coverage group type in APCD data (AP)}

\item{bsp_group_name}{Most frequently reported BSP group during requested date
range, case insensitive, can take multiple values (MD)}

\item{full_benefit_min}{Minimum time with full benefits during the requested
time period (percent scale) (MD)}

\item{cov_type}{Medicaid coverage type (FFS or MC) (MD)}

\item{mco_id}{Managed care organization ID (MD)}

\item{part_a_min}{Minimum time enrolled in Part A of Medicare during the requested 
time period (percent scale) (ME)}

\item{part_b_min}{Minimum time enrolled in Part B of Medicare during the requested 
time period (percent scale) (ME)}

\item{part_c_min}{Minimum time enrolled in Part C of Medicare during the requested 
time period (percent scale) (ME)}

\item{buy_in_min}{Minimum time with state buy in during the requested time
period (percent scale) (ME)}

\item{buy_in_max}{Maximum time with state buy in during the requested time
period (percent scale) (ME)}

\item{id}{Restrict to these specific APCD/Medicaid/Medicare IDs (A)}

\item{age_min}{Minimum age for cohort (integer), age is calculated as of 
last day of requested date range, defaults to 0 (A)}

\item{age_max}{Maximum age for cohort (integer), age is calculated as of 
last day of requested date range, defaults to 200 (A)}

\item{female}{Alone or in combination female gender over entire member history (A)}

\item{male}{Alone or in combination female gender over entire member history (A)}

\item{gender_me}{Most commonly reported gender, by time enrolled, case insensitive, 
can take multiple values (e.g., c("female", "multiple")) (A)}

\item{gender_recent}{Most recently reported gender, , case insensitive, 
can take multiple values (e.g., c("female", "multiple")) (AP/MD)}

\item{race_aian}{Alone or in combination American Indian/Alaska Native race over entire member history (MD/ME)}

\item{race_asian}{Alone or in combination Asian race over entire member history (MD/ME)}

\item{race_asian_pi}{Alone or in combination Asian/Pacific Islanfer race over entire member history (ME)}

\item{race_black}{Alone or in combination Black race over entire member history (MD/ME)}

\item{race_nhpi}{Alone or in combination Native Hawaiian/Pacific Islander race over entire member history (MD/ME)}

\item{race_latino}{Alone or in combination Latino race over entire member history (MD/ME)}

\item{race_white}{Alone or in combination white race over entire member history (MD/ME)}

\item{race_unk}{No recorded race over entire member history (MD/ME)}

\item{race_me}{Most frequently recorded race (excluding Latino) over entire member history,
case insensitive, can take multiple values (e.g., c("aian", "black")) (MD/ME)}

\item{race_eth_me}{Most frequently recorded race (including Latino) over entire member history,
case insensitive, can take multiple values (e.g., c("latino", "black")) (MD/ME)}

\item{race_recent}{Most recently recorded race (excluding Latino),
case insensitive, can take multiple values (e.g., c("aian", "black")) (MD/ME)}

\item{race_eth_recent}{Most recently recorded race (including Latino),
case insensitive, can take multiple values (e.g., c("latino", "black")) (MD/ME)}

\item{lang_amharic}{Alone or in combination Amharic written or spoken language over entire member history (MD)}

\item{lang_arabic}{Alone or in combination Arabic written or spoken language over entire member history (MD)}

\item{lang_chinese}{Alone or in combination Chinese written or spoken language over entire member history (MD)}

\item{lang_english}{Alone or in combination English written or spoken language over entire member history (MD)}

\item{lang_korean}{Alone or in combination Korean written or spoken language over entire member history (MD)}

\item{lang_russian}{Alone or in combination Russian written or spoken language over entire member history (MD)}

\item{lang_somali}{Alone or in combination Somali written or spoken language over entire member history (MD)}

\item{lang_spanish}{Alone or in combination Spanish written or spoken language over entire member history (MD)}

\item{lang_ukrainian}{Alone or in combination Ukrainian written or spoken language over entire member history (MD)}

\item{lang_vietnamese}{Alone or in combination Vietnamese written or spoken language over entire member history (MD)}

\item{lang_me}{Most frequently recorded spoken/written language over entire member history, 
case insensitive, can take multiple values (e.g., c("chinese", "english")) (MD)}

\item{lang_recent}{Most recently recorded spoken/written language, 
case insensitive, can take multiple values (e.g., c("chinese", "english")) (MD)}

\item{geo_zip}{Most frequently reported ZIP code during requested date range,
can take multiple values (e.g., c("98104", "98105")) (A)}

\item{geo_hra_code}{Most frequently reported health reporting area code during 
requested date range, can take multiple values (e.g., c("2100", "9000")) (MD)}

\item{geo_school_code}{Most frequently reported school district code during 
requested date range, can take multiple values (e.g., c("5307710", "5303540")) (MD)}

\item{geo_region}{Most frequently mapped HRA-based region during requested date range, 
(choose from east, north, seattle, south), case insensitive, can take multiple values (MD)}

\item{geo_county_code}{Most frequently reported county during requested date range 
(use FIPS codes), can take multiple values (AP/MD)}

\item{geo_ach_code}{Most recently reported accountable community of health 
during requested data range (use ACH codes), can take multiple values (AP)}

\item{geo_kc_ever}{Ever resided in King County (ME)}

\item{geo_kc_min}{Minimum amount of requested date range a person needs to have 
resided in King County to be included (AP/ME)}

\item{timevar_denom}{Which denominator is used to calculate the percentages
for time-varying parameters (e.g., dual_min, geo_kc_min). Choose from
duration (number of days in selected period, i.e., between from_date and to_date)
or cov_days (the number of days within selected period the person was actually
enrolled). Default is duration.}

\item{show_query}{Print the SQL query that is being run. Useful for debugging.
Default is TRUE}

\item{part_a_min}{Maximum time enrolled in Part A of Medicare during the requested 
time period (percent scale) (ME)}

\item{part_b_min}{Maximum time enrolled in Part B of Medicare during the requested 
time period (percent scale) (ME)}

\item{part_c_min}{Maximum time enrolled in Part C of Medicare during the requested 
time period (percent scale) (ME)}
}
\description{
\code{claims_elig} returns member eligibility and demographics.
}
\details{
LARGELY FOR INTERNAL USE
This function builds and sends a SQL query to return a member cohort with 
specified parameters, including coverage time period, coverage characteristics 
(e.g., Medicare dual eligibility), and member demographics. Can be used on three
data sources: APCD, Medicaid, and Medicare (also Medicaid/Medicare combined).
Parameters below have the following key to indicator which data source they 
can be used against: A = All sources, AP = APCD, MD = Medicaid, ME = Medicare
Most parameters default to NULL, which means that all values are included for
that field.
}
\examples{
\dontrun{
claims_elig(server = db.claims51, source = "apcd", 
  from_date = "2017-01-01", to_date = "2017-06-30")
claims_elig(server = db.claims51, source = "mcaid", 
  from_date = "2017-01-01", to_date = "2017-06-30", age_min = 18, 
  age_max = 64, lang_me = c("ARABIC", "SOMALI"), zip = c("98103", "98105"))
}

}
