% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elig_timevar_collapse.R
\name{elig_timevar_collapse}
\alias{elig_timevar_collapse}
\title{Collapse elig_timevar tables}
\usage{
elig_timevar_collapse(
  conn,
  server = c("phclaims", "hhsaw"),
  source = c("mcaid", "apcd"),
  dual = F,
  cov_time_day = T,
  last_run = F,
  ids = NULL,
  tpl = F,
  bsp_group_name = F,
  full_benefit = F,
  cov_type = F,
  mco_id = F,
  med_covgrp = F,
  pharm_covgrp = F,
  med_medicaid = F,
  med_medicare = F,
  med_commercial = F,
  pharm_medicaid = F,
  pharm_medicare = F,
  pharm_commercial = F,
  geo_add1 = F,
  geo_add2 = F,
  geo_city = F,
  geo_state = F,
  geo_zip = F,
  geocode_vars = F,
  geo_zip_code = F,
  geo_county = F,
  geo_ach = F
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package.}

\item{server}{Which server do you want to run the query against? NB. Currently only
Medicaid data is available on HHSAW.}

\item{source}{Which claims data source do you want to pull from?}

\item{dual}{Collapse over the dual eligiblity flag.}

\item{cov_time_day}{Recalculate coverage time in the new period. Default is TRUE.}

\item{last_run}{Bring in the last run date.}

\item{ids}{Restrict to specified IDs. Use format c("<id1>", "<id2>") or pass a vector.}

\item{tpl}{Collapse over the third party liability flag (Medicaid only).}

\item{bsp_group_name}{Collapse over the bsp_group_name field (Medicaid only)}

\item{full_benefit}{Collapse over the full_benefit field (Medicaid only).}

\item{cov_type}{Collapse over the cov_type field (Medicaid only).}

\item{mco_id}{Collapse over the mco_id field (Medicaid only)}

\item{med_covgrp}{Collapse over the med_covgrp field (APCD only).}

\item{pharm_covgrp}{Collapse over the pharm_covgrp field (APCD only).}

\item{med_medicaid}{Collapse over the med_medicaid field (APCD only).}

\item{med_medicare}{Collapse over the med_medicare field (APCD only).}

\item{med_commercial}{Collapse over the med_commercial field (APCD only).}

\item{pharm_medicaid}{Collapse over the pharm_medicaid field (APCD only).}

\item{pharm_medicare}{Collapse over the pharm_medicare field (APCD only).}

\item{pharm_commercial}{Collapse over the pharm_commercial field (APCD only).}

\item{geo_add1}{Collapse over the geo_add1 field (Medicaid only).}

\item{geo_add2}{Collapse over the geo_add2 field (Medicaid only).}

\item{geo_city}{Collapse over the geo_city field (Medicaid only).}

\item{geo_state}{Collapse over the geo_state field (Medicaid only).}

\item{geo_zip}{Collapse over the geo_zip field (Medicaid only).}

\item{geocode_vars}{Bring in all geocded data elements (geo_zip_centroid, 
geo_street_centroid, geo_county_code, geo_tractce10, geo_hra_code, 
geo_school_code). Default is FALSE.}

\item{geo_zip_code}{Collapse over the geo_zip_code field (APCD only).}

\item{geo_county}{Collapse over the geo_countyfield (APCD only).}

\item{geo_ach}{Collapse over the geo_ach field (APCD only).}
}
\description{
\code{elig_timevar_collapse} collapses elig_timevar tables
}
\details{
Standard time-varying eligibility (elig_timevar) tables have new lines 
(i.e., new from_date and to_date values) for a change in any time-varying element. 
This function allows users to generate a new time-varying eligibility
table (elig_timevar) that creates a new line only for desired data elements. 
All fields that define what the new table is collapsed over are set to FALSE 
by default. NB. Function does not yet support Medicare or combined Medicaid/Medicare
tables.
}
\examples{
\dontrun{
new_timevar <- elig_timevar_collapse(conn = db_claims51, source = "mcaid",
full_benefit = T, geo_add1 = T, geo_city = T, geo_zip = T, geocode_vars = T)
new_timevar2 <- elig_timevar_collapse(conn = db_claims51, source = "apcd",
ids = c("123", "456", "789"), med_covgrp = T, geo_county = T)
}

}
