% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_var}
\alias{list_var}
\title{Create a list of quosures from unquoted variable names}
\usage{
list_var(...)
}
\arguments{
\item{...}{Variables that will be passed as a list to another function}
}
\value{
A list of quosures created from the input variables
}
\description{
Accepts a list of unquoted variable names and returns a list of quosures
for passing to any function that expects quosures.
}
\details{
This is a utility function commonly used in conjunction with dplyr and 
other tidyverse functions that expect quosures as input. The function uses 
`rlang::quos()` internally to capture the expressions.
}
\note{
This function uses a local variable named 'list' which shadows R's built-in 
list() function. This may not cause functional issues, but it is a non-standard 
coding practice and may make debugging more annoying.
}
\examples{
\dontrun{
# Create quosures for use with dplyr functions
vars_to_select <- list_var(name, age, income)

# Use with select functions
selected_data <- dplyr::select(data, !!!vars_to_select)
}

}
\keyword{utilities}
