% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elig_cohort.R
\name{mcaid_cohort_f}
\alias{mcaid_cohort_f}
\title{Medicaid member cohort}
\usage{
mcaid_cohort_f(from_date = Sys.Date() - months(12), to_date = Sys.Date() -
  months(6), covmin = 0, ccov_min = 1, covgap_max = "null",
  dualmax = 100, agemin = 0, agemax = 200, female = "null",
  male = "null", aian = "null", asian = "null", black = "null",
  nhpi = "null", white = "null", latino = "null", zip = "null",
  zregion = "null", english = "null", spanish = "null",
  vietnamese = "null", chinese = "null", somali = "null",
  russian = "null", arabic = "null", korean = "null",
  ukrainian = "null", amharic = "null", maxlang = "null", id = "null")
}
\arguments{
\item{from_date}{Begin date for Medicaid coverage period, "YYYY-MM-DD", defaults to 12 months prior to today's date}

\item{to_date}{End date for Medicaid coverage period, "YYYY-MM-DD", defaults to 6 months prior to today's date}

\item{covmin}{Minimum coverage required during requested date range (percent scale), defaults to 0}

\item{ccov_min}{Minimum continuous coverage required during requested date range (days), defaults to 1}

\item{covgap_max}{Maximum gap in continuous coverage allowed during requested date range (days), defaults to null}

\item{dualmax}{Maximum Medicare-Medicaid dual eligibility coverage allowed during requested date range (percent scale), defaults to 100}

\item{agemin}{Minimum age for cohort (integer), age is calculated as of last day of requested date range, defaults to 0}

\item{agemax}{Maximum age for cohort (integer), age is calculated as of last day of requested date range, defaults to 200}

\item{female}{Alone or in combination female gender over entire member history, defaults to null}

\item{male}{Alone or in combination female gender over entire member history, defaults to null}

\item{aian}{Alone or in combination American Indian/Alaska Native race over entire member history, defaults to null}

\item{asian}{Alone or in combination Asian race over entire member history, defaults to null}

\item{black}{Alone or in combination Black race over entire member history, defaults to null}

\item{nhpi}{Alone or in combination Native Hawaiian/Pacific Islander race over entire member history, defaults to null}

\item{white}{Alone or in combination white race over entire member history, defaults to null}

\item{latino}{Alone or in combination Latino race over entire member history, defaults to null}

\item{zip}{Most frequently reported ZIP code during requested date range, eg. "98103,98105", defaults to null}

\item{zregion}{Most frequently mapped ZIP code-based region during requested date range, e.g, "east,north,seattle,south", defaults to null}

\item{english}{Alone or in combination English written or spoken language over entire member history, defaults to null}

\item{spanish}{Alone or in combination Spanish written or spoken language over entire member history, defaults to null}

\item{vietnamese}{Alone or in combination Vietnamese written or spoken language over entire member history, defaults to null}

\item{chinese}{Alone or in combination Chinese written or spoken language over entire member history, defaults to null}

\item{somali}{Alone or in combination Somali written or spoken language over entire member history, defaults to null}

\item{russian}{Alone or in combination Russian written or spoken language over entire member history, defaults to null}

\item{arabic}{Alone or in combination Arabic written or spoken language over entire member history, defaults to null}

\item{korean}{Alone or in combination Korean written or spoken language over entire member history, defaults to null}

\item{ukrainian}{Alone or in combination Ukrainian written or spoken language over entire member history, defaults to null}

\item{amharic}{Alone or in combination Amharic written or spoken language over entire member history, defaults to null}

\item{maxlang}{Most frequently reported spoken/written language, e.g. "SOMALI,ARABIC", defaults to null}

\item{id}{List of requested Medicaid ProviderOne IDs, defaults to null}
}
\description{
\code{mcaid_cohort_f} builds a SQL query to return a Medicaid member cohort.
}
\details{
LARGELY FOR INTERNAL USE
This function builds and sends a SQL query to return a Medicaid member cohort with specified parameters, including
coverage time period, coverage characteristics (e.g. Medicare dual eligibility), and member demographics.
}
\examples{
\dontrun{
mcaid_cohort_f(from_date = "2017-01-01", to_date = "2017-06-30")
mcaid_cohort_f(from_date = "2017-01-01", to_date = "2017-06-30", agemin = 18, 
agemax = 64, maxlang = "ARABIC,SOMALI", zip = "98103,98105")  
}

}
