% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid_condition.R
\name{mcaid_condition_f}
\alias{mcaid_condition_f}
\title{Medicaid member chronic health condition status}
\usage{
mcaid_condition_f(
  server,
  condition,
  from_date,
  to_date,
  join_type = "none",
  df_join_name
)
}
\arguments{
\item{server}{SQL server connection created using \code{odbc} package}

\item{condition}{The chronic health condition requested from SQL Server, current valid values include "asthma", "chr_kidney_dis", 
"copd", "depression", "diabetes", "heart_failure", "hypertension", and "ischemic_heart_dis"}

\item{from_date}{Begin date for Medicaid coverage period, "YYYY-MM-DD", defaults to 12 months prior to today's date}

\item{to_date}{End date for Medicaid coverage period, "YYYY-MM-DD", defaults to 6 months prior to today's date}

\item{join_type}{Type of requested join - left, right or inner, defaults to no join.}
}
\description{
\code{mcaid_condition_f} builds a SQL query to return Medicaid member chronic health condition information.
}
\details{
LARGELY FOR INTERNAL USE
This function builds and sends a SQL query to return a Medicaid member cohort with a specified chronic health condition.
If requested, the function will also join the returned data to a specified data frame in R, joining on Medicaid member ID. 
Users can specify the join type (left, right, inner). By default "ever" status is returned - for example a request for diabetic
members will return members with any history of diabetes in the Medicaid claims database, using the Chronic Conditions Warehouse
definition from CDC. If a date range is supplied, the function will only return members who were identified as having the
condition during the date range (a function of both Medicaid coverage and health care diagnostic information).
}
\examples{
\dontrun{
condition <- mcaid_condition_f(server = db.claims51, condition = "chr_kidney_dis")
condition <- mcaid_condition_f(server = db.claims51, condition = "asthma", join_type = "inner", df_join_name = elig_test)
condition <- mcaid_condition_f(server = db.claims51, condition = "depression", from_date = "2017-01-01", to_date = "2017-01-31")
}

}
