% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sqlbatch_f}
\alias{sqlbatch_f}
\title{Execute batched SQL statements with final result return}
\usage{
sqlbatch_f(server, sqlbatch)
}
\arguments{
\item{server}{SQL server connection created using `odbc` package}

\item{sqlbatch}{A list of SQL queries where all but the last are executed 
without returning results. The final query should return the desired result set.}
}
\value{
A data frame containing the results of the final SQL statement
}
\description{
Prepares and sends batched SQL statements to SQL Server with all but the 
final statement executed without returning results, and the final statement 
returning a result set.
}
\details{
This function is useful for executing setup queries (like creating temporary 
tables or updating records) followed by a final SELECT statement. All queries except 
the last are executed without returning data to improve performance. The connection 
should be established using the `odbc` package.
}
\examples{
\dontrun{
# Establish connection
conn <- odbc::dbConnect(odbc::odbc(), "MyServer")

# Execute setup queries followed by a SELECT statement
queries <- list(
  "CREATE TEMP TABLE #temp AS SELECT * FROM main_table WHERE year = 2025",
  "UPDATE #temp SET status = 'processed'",
  "SELECT * FROM #temp"
)
result <- sqlbatch_f(conn, queries)
}

}
\keyword{database}
\keyword{utilities}
