% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_causes.R
\name{top_causes}
\alias{top_causes}
\title{Top N conditions seen among Mediciad patients}
\usage{
top_causes(
  conn,
  server = c("hhsaw"),
  source = c("apcd", "mcaid", "mcaid_mcare", "mcare"),
  cohort,
  cohort_id = NULL,
  renew_ids = T,
  from_date = NULL,
  to_date = NULL,
  ind_dates = F,
  ind_from_date = NULL,
  ind_to_date = NULL,
  top = 15,
  catch_all = F,
  primary_dx = T,
  type = c("ed", "inpatient", "all"),
  override_all = F
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package.}

\item{server}{Which server do you want to run the query against? NB. Currently just HHSAW.}

\item{source}{Which claims data source do you want to pull from?}

\item{cohort}{The group of individuals of interest. Note: it is possible to generate a cohort on the fly
using \code{\link{claims_elig}}.}

\item{cohort_id}{The field that contains the ID in the cohort data. Defaults to id_apde.}

\item{renew_ids}{Option to avoid reloading ID fields to temp table.}

\item{from_date}{Begin date for claims period, "YYYY-MM-DD", defaults to start of 
the previous calendar year.}

\item{to_date}{End date for claims period, "YYYY-MM-DD", defaults to end of the previous calendar year
or 6 months prior to today's date, whichever is earlier.}

\item{ind_dates}{Flag to indicate that individualized dates are used to narrow
the default date window.}

\item{ind_from_date}{Field in the cohort data that contains an individual from date.}

\item{ind_to_date}{Field in the cohort data that contains an individual to date.}

\item{top}{The maximum number of condition groups that will be returned, default is 15.}

\item{catch_all}{Determines whether or not catch_all codes are included in the list,
default is no.}

\item{primary_dx}{Whether or not to only look at the primary diagnosis field, default is TRUE.}

\item{type}{Which types of visits to include. Choose from the following:
ed (any ED visit), 
inpatient (any inpatient visit)
all (all claims, must be paired with override_all option)}

\item{override_all}{Override the warning message about pulling all claims, default is FALSE.}
}
\description{
\code{top_causes} identifies the top N causes for a given set of visits.
}
\details{
This function builds a temp table with the IDs of a cohort of interest.
It then creates a SQL query to find claims for that cohort made in a given time frame
The top N categories are selected and a count of the claims in each category given.
There are optional flags for the following:
1) To limit categories to just the primary dx vs. all dx.
2) To restrict to certain visit types (e.g., ED visits, hospitalizations)
}
\examples{
\dontrun{
system.time(mcaid_only <- claims_elig(conn = db_hhsaw, source = "mcaid", server = "hhsaw",
    from_date = "2014-01-01", to_date = "2015-02-25",
    geo_zip = c("98104", "98133", "98155"),
    cov_type = "FFS", race_asian = 1, 
    show_query = T))
top_15_dynamic <- top_causes(
    cohort = mcaid_only, top = 3, conn = db_hhsaw, source = "mcaid",
    server="hhsaw"
)
}

}
