% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid_top_n_conds.R
\name{top_causes_f}
\alias{top_causes_f}
\title{Top N conditions seen among Mediciad patients}
\usage{
top_causes_f(cohort, cohort_id = NULL, server = db.claims51,
  from_date = NULL, to_date = NULL, top = 15, primary_dx = T,
  ed_all = T, ed_avoid_ny = T, ed_avoid_ca = T, inpatient = T)
}
\arguments{
\item{cohort}{The group of individuals of interest. Note: it is possible to generate a cohort on the fly
using \code{\link{mcaid_elig_f}}.}

\item{server}{SQL server connection created using \code{odbc} package.}

\item{from_date}{Begin date for claims period, "YYYY-MM-DD", defaults to start of 
the previous calendar year.}

\item{to_date}{End date for claims period, "YYYY-MM-DD", defaults to end of the previous calendar year
or 6 months prior to today's date, whichever is earlier.}

\item{top}{The maximum number of condition groups that will be returned, default is 15.}

\item{primary_dx}{Whether or not to only look at the primary diagnosis field, default is TRUE.}

\item{ed_all}{Will include any ED visit}

\item{ed_avoid_ny}{Will include any avoidable ED visit (based on NYU classification)}

\item{ed_avoid_ca}{Will include any avoidable ED visit (based on CA classification)}

\item{inpatient}{Will include any inpatient visit}

\item{cohort_d}{The field that contains the Medicaid ID in the cohort data. Defaults to id.}
}
\description{
\code{top_causes_f} identifies the top N causes for a given set of visits.
}
\details{
This function builds a temp table with the IDs of a cohort of interest.
It then creates a SQL query to find claims for that cohort made in a given time frame
The top N categories are selected and a count of the claims in each category given.
There are optional flags for the following:
1) To limit categories to just the primary dx vs. all dx.
2) To restrict to certain visit types (e.g., ED visits, hospitalizations)
}
\examples{
\dontrun{
top_15 <- top_causes_f(cohort = focus_pop, cohort_id = id, server = db.claims51)
top_15_dynamic <- top_causes_f(cohort = mcaid_elig_f(server = db.claims51, 
from_date = "2017-01-01", to_date = "2017-12-31", korean = 1, zip = "98103"), top = 3)
}

}
