% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper}
\alias{helper}
\alias{list_var}
\title{Broad use helper functions}
\usage{
list_var(...)
}
\arguments{
\item{...}{Variables that will be passed as a list to another function}
}
\description{
Various functions used to support processing and analysis
of Medicaid claims data.

\code{list_var} accepts a list of unquoted variable names and returns a list of quosures
for passing to any function.
}
