% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper}
\alias{helper}
\alias{list_var}
\alias{sqlbatch_f}
\alias{suppress_f}
\title{Broad use helper functions}
\usage{
list_var(...)

sqlbatch_f(server, sqlbatch)

suppress_f(df, lower = 1, upper = 9, varlist)
}
\arguments{
\item{...}{Variables that will be passed as a list to another function}

\item{server}{SQL server connection created using \code{odbc} package}

\item{sqlbatch}{Any number of SQL queries in list format}

\item{df}{Data frame}

\item{lower}{Lower bound for small number suppression, defaults to 1}

\item{upper}{Upper bound for small number suppression, defaults to 9}

\item{varlist}{Vector of variable names to suppress in quotes (e.g. c = ("count", "mean"))}
}
\description{
Various functions used to support processing and analysis
of Medicaid claims data.

\code{list_var} accepts a list of unquoted variable names and returns a list of quosures
for passing to any function.

\code{sqlbatch_f} prepares and sends batched SQL statements to SQL Server with the final statement
returning a result set.
}
