% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabloop.R
\name{tabloop_f}
\alias{tabloop_f}
\title{Tabulation loop

\code{tabloop_f} tabulates a data frame (e.g. summarize ndistinct) over fixed and 
looped by variables, binding all output as a single data frame.}
\usage{
tabloop_f(df, unit, loop, fixed = NULL)
}
\arguments{
\item{df}{A data frame in tidy format}

\item{unit}{A variable that identifies the unit of tabulation (e.g.person, row, etc.)}

\item{loop}{A list of the loop by variables, requires use of \code{list_var}, required}

\item{fixed}{A list of the fixed by variables, requires use of \code{list_var}, defaults to null}
}
\description{
This function tabulates a single data frame over fixed and looped by variables, and binds
all output as a single data frame. Fixed by variables are variables by which the data frame will be disaggregated
for all loop variables, whereas loop variables will only be disaggregated separately. For example, a combination of
region for fixed and age group and sex for loop would produce counts by age group and sex for each region, but not
counts for each sex by age group.

The function accepts a row ID variable and summarizes distinct counts of this variable.
The function will produce zero counts for all by variable values that exist in the full join of the fixed
and loop by variable matrix.
}
\examples{
\dontrun{
tabloop_f(df = mcaid_cohort, unit = id, loop = list_var(gender, race), fixed = list_var(region))
tabloop_f(df = mcaid_cohort, unit = id, loop = list_var(gender, race, zip_code, cov_grp, language))
}

}
